move code to file, rename file
